[SYCL] update current device tests
