+++ tests/extension/oneapi_current_device/CMakeLists.txt
@@ -1,4 +1,4 @@
-if(SYCL_CTS_ENABLE_EXT_ONEAPI_ROOT_GROUP_TESTS)
+if(SYCL_CTS_ENABLE_EXT_ONEAPI_CURRENT_DEVICE_TESTS)
     file(GLOB test_cases_list *.cpp)
 
     add_cts_test(${test_cases_list})
